Initial consumer benchmark implementation

+++ src/producer/content.rs
@@ -3,13 +3,13 @@ use rand::{self, Rng};
 pub struct CachedMessages(Vec<Vec<u8>>);
 
 impl CachedMessages {
-    pub fn new(msg_size: usize, cache_size: usize) -> CachedMessages {
+    pub fn new(msg_size: u64, cache_size: u64) -> CachedMessages {
         let messages = (0..(cache_size / msg_size))
             .map(|_| {
                 rand::thread_rng()
                     .gen_iter::<u8>()
                     .map(|v| v % 86 + 40)
-                    .take(msg_size)
+                    .take(msg_size as usize)
                     .collect::<Vec<u8>>()
             })
             .collect::<Vec<_>>();
